refactor: move RegisterScreenEvent to separate file
